knowledge-well/requirements.txt
# --- API ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# --- HTTP client & env ---
httpx>=0.27.0
python-dotenv>=1.0.1


# --- Vector DB (local) ---
chromadb>=0.5.0,<0.6

# --- Embeddings ---
sentence-transformers>=2.7.0

# PyTorch (CPU by default). If you want GPU, see notes below.
torch>=2.3.0,<2.5

# (Optional) Useful utils
tqdm>=4.66.0

google-generativeai>=0.5.0,<1.0.0

sentence-transformers

openai

tqdm>=4.66.0

anyio==4.4.0


